Add an uuid field for the play model and API
Ansible generates a unique uuid for every play and this uuid is
supplied to the callbacks as well as the action plugins.

This will allow action plugins to find the play they are running
in. Since the play references the playbook that is running, it will
allow action plugins to positively identify the playbook by querying
the API for the play by searching for the uuid.

Change-Id: I599ea9150649db5eb5d2eaab503b47a58ea27c4f